% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ames_scores_results}
\alias{ames_scores_results}
\title{Ames exampled score results}
\value{
A data frame of exampled score results.
}
\description{
This data an exampled score results for ames. See
\code{\link[=show_best_desirability_prop]{show_best_desirability_prop()}} for how it is used
for demonstration, and see \code{\link[=fill_safe_values]{fill_safe_values()}}
for how it is created.
}
\examples{
data(ames_scores_results)
}
