% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{arrange_score}
\alias{arrange_score}
\title{Arrange score result \code{results}}
\usage{
arrange_score(x, ...)
}
\arguments{
\item{x}{NULL}

\item{...}{NULL}
}
\description{
Arrange score result \code{results}
}
