% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as_result_obj}
\alias{as_result_obj}
\title{Assign class result_obj to score object score_obj}
\usage{
as_result_obj(x, ...)
}
\arguments{
\item{x}{NULL}

\item{...}{NULL}
}
\description{
Assign class result_obj to score object score_obj
}
