% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{attach_score}
\alias{attach_score}
\title{Attach score \code{res} to score object \code{score_obj}}
\usage{
attach_score(x, ...)
}
\arguments{
\item{x}{NULL}

\item{...}{NULL}
}
\description{
Attach score \code{res} to score object \code{score_obj}
}
