% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bind_scores}
\alias{bind_scores}
\title{Bind all metadata \code{score_obj} and score result \code{results}, and assign class \code{score_set} to combined scores.}
\usage{
bind_scores(x, ...)
}
\arguments{
\item{x}{A list where each element is a score object of class \code{score_obj}.}
}
\description{
Bind all metadata \code{score_obj} and score result \code{results}, and assign class \code{score_set} to combined scores.
}
