% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_cutoff}
\alias{check_cutoff}
\title{Check if cutoff value \code{cutoff} is specified}
\usage{
check_cutoff(cutoff)
}
\arguments{
\item{cutoff}{NULL}
}
\description{
Check if cutoff value \code{cutoff} is specified
}
