% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_num_terms}
\alias{check_num_terms}
\title{Check if number of predictors \code{num_terms} is specified}
\usage{
check_num_terms(num_terms)
}
\arguments{
\item{num_terms}{NULL}
}
\description{
Check if number of predictors \code{num_terms} is specified
}
