% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_prop_terms}
\alias{check_prop_terms}
\title{Check if proportion of predictors \code{prop_terms} is specified}
\usage{
check_prop_terms(prop_terms)
}
\arguments{
\item{prop_terms}{NULL}
}
\description{
Check if proportion of predictors \code{prop_terms} is specified
}
