% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_target}
\alias{check_target}
\title{Check if target value \code{target} is supplied}
\usage{
check_target(target)
}
\arguments{
\item{target}{NULL}
}
\description{
Check if target value \code{target} is supplied
}
