% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{class_score_list}
\alias{class_score_list}
\title{Construct an S7 subclass of base R's \code{list} for Method Dispatch}
\format{
An object of class \code{S7_S3_class} of length 3.
}
\usage{
class_score_list
}
\description{
\code{class_score_list} is an S7 subclass of S3 base R's \code{list}, used for method dispatch in
\code{\link[=bind_scores]{bind_scores()}} and \code{\link[=fill_safe_values]{fill_safe_values()}}.
}
\keyword{datasets}
