% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dont_log_pvalues}
\alias{dont_log_pvalues}
\title{Disable -log10 transformation of p-values}
\usage{
dont_log_pvalues(x)
}
\arguments{
\item{x}{A score class object.}
}
\value{
The modified score class object with \code{neg_log10} set to \code{FALSE}.
}
\description{
Disable -log10 transformation of p-values
}
