% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{enable_fdr()}
\alias{enable_fdr()}
\alias{enable_fdr}
\title{Enable the fdr (false discovery rate) adjusted p-values, using the Benjamini-Hochberg (BH) procedure}
\usage{
enable_fdr(x)
}
\arguments{
\item{x}{A score class object.}
}
\value{
The modified score class object with \code{neg_log10} set to \code{FALSE}.
}
\description{
Enable the fdr (false discovery rate) adjusted p-values, using the Benjamini-Hochberg (BH) procedure
}
