% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fill_safe_value}
\alias{fill_safe_value}
\title{Fill safe value}
\usage{
fill_safe_value(x, ...)
}
\arguments{
\item{x}{A score class object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Fill safe value
}
