% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filter_score_auto}
\alias{filter_score_auto}
\title{Filter score result \code{results} based on number or proportion of predictors with
optional cutoff value}
\usage{
filter_score_auto(x, ...)
}
\arguments{
\item{x}{NULL}

\item{...}{NULL}
}
\description{
Filter score result \code{results} based on number or proportion of predictors with
optional cutoff value
}
