% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filter_score_num}
\alias{filter_score_num}
\title{Filter score \code{res} based on number of predictors}
\usage{
filter_score_num(x, ...)
}
\arguments{
\item{x}{NULL}

\item{...}{NULL}
}
\description{
Filter score \code{res} based on number of predictors
}
