% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filter_score_prop}
\alias{filter_score_prop}
\title{Filter score based on proportion of predictors}
\usage{
filter_score_prop(x, ...)
}
\arguments{
\item{x}{NULL}

\item{...}{NULL}
}
\description{
Filter score based on proportion of predictors
}
