% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters-package.R
\docType{package}
\name{filters-package}
\alias{filters}
\alias{filters-package}
\title{filters: Tools for Feature Selection Using Supervised Filter-Based Techniques}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/filters}
  \item \url{https://tidymodels.github.io/filters/}
  \item Report bugs at \url{https://github.com/tidymodels/filters/issues}
}

}
\author{
\strong{Maintainer}: Frances Lin \email{franceslinyc@gmail.com}

Authors:
\itemize{
  \item Max Kuhn \email{max@posit.co}
  \item Emil Hvitfeldt \email{emil.hvitfeldt@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
