% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtro-package.R
\docType{package}
\name{filtro-package}
\alias{filtro}
\alias{filtro-package}
\title{filtro: Tools for Feature Selection Using Supervised Filter-Based Methods}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tidymodels.github.io/filtro}
  \item \url{https://github.com/tidymodels/filtro}
  \item Report bugs at \url{https://github.com/tidymodels/filtro/issues}
}

}
\author{
\strong{Maintainer}: Frances Lin \email{franceslinyc@gmail.com}

Authors:
\itemize{
  \item Max Kuhn \email{max@posit.co}
  \item Emil Hvitfeldt \email{emil.hvitfeldt@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
