% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_score.R
\name{fit}
\alias{fit}
\title{A generic for fitting score objects.}
\usage{
fit(object, ...)
}
\arguments{
\item{object}{A score class object (e.g., \code{score_cor_pearson}).}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
A generic for fitting score objects.
}
\details{
For information \code{fit()} methods for specific score objects, see the "Details"
section (and the "Estimating the scores" subsection) for the score object
(e.g., \code{?score_cor_pearson}). There may be other arguments to fit methods
that are not documented here.
}
