% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-aov.R
\name{fit_class_score_aov}
\alias{fit_class_score_aov}
\title{Compute ANOVA p-values scores}
\usage{
fit_class_score_aov(object, formula, data, ...)
}
\arguments{
\item{object}{A score class object based on \code{class_score_aov}.}

\item{formula}{A standard R formula with a single outcome on the right-hand
side and one or more predictors (or \code{.}) on the left-hand side. The data are
processed via \code{\link[stats:model.frame]{stats::model.frame()}}.}

\item{data}{A data frame containing the relevant columns defined by the
formula.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Compute ANOVA p-values scores
}
