% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-cor.R
\name{get_scores_cor}
\alias{get_scores_cor}
\title{Compute Pearson or Spearman correlation coefficients}
\usage{
get_scores_cor(score_obj, data, outcome)
}
\arguments{
\item{score_obj}{NULL}

\item{data}{NULL}

\item{outcome}{NULL}
}
\description{
Compute Pearson or Spearman correlation coefficients
}
