% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-cross_tab.R
\name{get_scores_cross_tab}
\alias{get_scores_cross_tab}
\title{Compute cross tabulation p-value scores using Fisher’s exact test and the chi-squared test}
\usage{
get_scores_cross_tab(score_obj, data, outcome, ...)
}
\arguments{
\item{score_obj}{NULL}

\item{data}{NULL}

\item{outcome}{NULL}

\item{...}{NULL}
}
\description{
Compute cross tabulation p-value scores using Fisher’s exact test and the chi-squared test
}
