% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-forest_imp.R
\name{get_scores_forest_importance}
\alias{get_scores_forest_importance}
\title{Compute feature importance scores using a random forest, a conditional random forest, or
an oblique random forest}
\usage{
get_scores_forest_importance(score_obj, data, outcome, ...)
}
\arguments{
\item{score_obj}{NULL}

\item{data}{NULL}

\item{outcome}{NULL}

\item{...}{NULL}
}
\description{
Compute feature importance scores using a random forest, a conditional random forest, or
an oblique random forest
}
