% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-info_gain.R
\name{get_scores_info_gain}
\alias{get_scores_info_gain}
\title{Compute information gain}
\usage{
get_scores_info_gain(score_obj, data, outcome, ...)
}
\arguments{
\item{score_obj}{NULL}

\item{data}{NULL}

\item{outcome}{NULL}

\item{...}{NULL}
}
\description{
Compute information gain
}
