% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-roc_auc.R
\name{get_scores_roc_auc}
\alias{get_scores_roc_auc}
\title{Compute area under the Receiver Operating Characteristic curve (ROC AUC)}
\usage{
get_scores_roc_auc(score_obj, data, outcome)
}
\arguments{
\item{score_obj}{NULL}

\item{data}{NULL}

\item{outcome}{NULL}
}
\description{
Compute area under the Receiver Operating Characteristic curve (ROC AUC)
}
