% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_difference.R
\name{mean_difference}
\alias{mean_difference}
\title{Calculate mean difference}
\usage{
mean_difference(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A single numeric value.
}
\description{
Calculate mean difference
}
\examples{
x <- c(1, 2, 3)
y <- c(4, 5, 6)
mean_difference(x, y)
}
