% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{new_score_obj}
\alias{new_score_obj}
\title{Tools for creating new score object}
\usage{
new_score_obj(
  subclass = c("cat_num", "cat_cat", "num_num", "any"),
  outcome_type = c("numeric", "factor"),
  predictor_type = c("numeric", "factor"),
  case_weights = NULL,
  range = NULL,
  inclusive = NULL,
  fallback_value = NULL,
  score_type = NULL,
  trans = NULL,
  sorts = NULL,
  direction = NULL,
  deterministic = NULL,
  tuning = NULL,
  ties = NULL,
  calculating_fn = NULL,
  label = NULL,
  ...
)
}
\arguments{
\item{subclass}{A character string.}

\item{outcome_type}{A character string. One of:
\itemize{
\item \code{"numeric"}
\item \code{"factor"}
}}

\item{predictor_type}{A character string. One of:
\itemize{
\item \code{"numeric"}
\item \code{"factor"}
}}

\item{case_weights}{NULL}

\item{range}{A numeric vector of length two, specifying the minimum and maximum
possible values, respectively.}

\item{inclusive}{A logical vector of length two, indicating whether the lower and
upper bounds of the range are inclusive (\code{TRUE}) or exclusive (\code{FALSE}).}

\item{fallback_value}{A numeric scalar used as a fallback value. Typical values
include:
\itemize{
\item \code{0}
\item \code{1}
\item \code{Inf}
}}

\item{score_type}{A character string indicating the type of scoring metric to compute.
Available options include:
\itemize{
\item ANOVA F-Test: \code{"fstat"}, \code{"pval"}
\item Correlation: \code{"pearson"}, \code{"spearman"}
\item Cross Tabulation: \code{"pval_chisq"}, \code{"pval_fisher"}
\item Random Forest:\code{"imp_rf"}, \code{"imp_rf_conditional"}, \code{"imp_rf_oblique"}
\item Information Gain: \code{"infogain"}, \code{"gainratio"}, \code{"symuncert"}
\item ROC AUC: \code{"roc_auc"}
}}

\item{trans}{A \code{trans} object from the \pkg{scales} package, such as
\code{\link[scales:transform_log]{scales::transform_log10()}} or \code{\link[scales:transform_reciprocal]{scales::transform_reciprocal()}}. Or use built-in
functions, such as \code{\link[=transform_abs]{transform_abs()}} or \code{\link[=transform_neg_log10]{transform_neg_log10()}}.
Create custom transforms with \code{\link[scales:new_transform]{scales::trans_new()}}.}

\item{sorts}{A character string indicating used to sort the scores. Common options include:
\itemize{
\item \code{identity}
\item \code{abs}
\item \code{function(score) max(score, 1 - score)}
}}

\item{direction}{A character string indicating the optimization direction. One of:
\itemize{
\item \code{"maximize"}
\item \code{"minimize"}
\item \code{"target"}
}}

\item{deterministic}{A logical value, indicating whether the resulting score is
deterministic (\code{TRUE}) or random (\code{FALSE}).}

\item{tuning}{A logical value, indicating whether the model should be tuned
(\code{TRUE}) or not (\code{FALSE}).}

\item{ties}{A logical value indicating whether ties in score can occur (\code{TRUE})
or not (\code{FALSE}).}

\item{calculating_fn}{NULL}

\item{label}{A named named character string.}

\item{...}{NULL}
}
\description{
Tools for creating new score object
}
\examples{
NULL
}
