% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rank_best_score_dense}
\alias{rank_best_score_dense}
\title{Rank score based on \code{dplyr::dense_rank()}, where tied values receive the
same rank and ranks are without gaps \emph{(singular)}}
\usage{
rank_best_score_dense(x, ...)
}
\arguments{
\item{x}{A score class object, i.e., \verb{score_*}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A tibble of score results.
}
\description{
Rank score based on \code{dplyr::dense_rank()}, where tied values receive the
same rank and ranks are without gaps \emph{(singular)}
}
