% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rank_best_score_min}
\alias{rank_best_score_min}
\title{Rank score based on \code{dplyr::min_rank()}, where tied values receive the
same rank and ranks are with gaps \emph{(singular)}}
\usage{
rank_best_score_min(x, ...)
}
\arguments{
\item{x}{A score class object, i.e., \verb{score_*}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A tibble of score results.
}
\description{
Rank score based on \code{dplyr::min_rank()}, where tied values receive the
same rank and ranks are with gaps \emph{(singular)}
}
