% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rank_score_dense}
\alias{rank_score_dense}
\title{Rank score based on dense_rank (Need a better title)}
\usage{
rank_score_dense(x, ...)
}
\arguments{
\item{x}{NULL}

\item{...}{NULL}
}
\description{
Rank score based on dense_rank (Need a better title)
}
