% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rank_score_min}
\alias{rank_score_min}
\title{Rank score \code{res} based on min_rank (Need a better title)}
\usage{
rank_score_min(x, ...)
}
\arguments{
\item{x}{NULL}

\item{...}{NULL}
}
\description{
Rank score \code{res} based on min_rank (Need a better title)
}
