% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_score.R
\name{required_pkgs}
\alias{required_pkgs}
\title{An api mimicking \code{generics::required_pkgs()} to return the \code{packages} property}
\usage{
required_pkgs(x, ...)
}
\arguments{
\item{x}{A score class object (e.g., \code{score_cor_pearson}).}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
A generic for determine which packages are required for operation.
}
