% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-aov.R
\name{score_aov}
\alias{score_aov}
\title{Title}
\usage{
score_aov(
  range = c(0, Inf),
  trans = NULL,
  score_type = "fstat",
  direction = "maximize"
)
}
\arguments{
\item{range}{NULL}

\item{trans}{NULL}

\item{score_type}{NULL}

\item{direction}{NULL}
}
\description{
Title
}
\examples{
NULL
}
