% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-aov.R
\docType{data}
\name{score_aov_pval}
\alias{score_aov_pval}
\alias{score_aov_fstat}
\title{Scoring via analysis of variance hypothesis tests}
\format{
An object of class \code{filtro::class_score_aov} (inherits from \code{filtro::class_score}, \code{S7_object}) of length 1.

An object of class \code{filtro::class_score_aov} (inherits from \code{filtro::class_score}, \code{S7_object}) of length 1.
}
\usage{
score_aov_pval

score_aov_fstat
}
\arguments{
\item{object}{A score class object based on \code{class_score_aov}.}

\item{formula}{A standard R formula with a single outcome on the right-hand
side and one or more predictors (or \code{.}) on the left-hand side. The data are
processed via \code{\link[stats:model.frame]{stats::model.frame()}}.}

\item{data}{A data frame containing the relevant columns defined by the
formula.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
These objects are used when either:
\itemize{
\item The predictors are numeric and the outcome is a factor/category, or
\item The predictors are factors and the outcome is numeric.
}

In either case, a linear model (via \code{\link[stats:lm]{stats::lm()}}) is created with the proper
variable roles, and the overall p-value for the hypothesis that all means are
equal is computed via the standard F-statistic. The p-value that is returned
is transformed to be \code{-log10(p_value)} so that larger values are associated
with more important predictors.

\code{score_aov_pval} and \code{score_aov_fstat} are objects that define the technique.
To apply the filter on data, you would use the \code{\link[=fit]{fit()}} method:

\preformatted{
  fit(score_aov_pval, formula, data)
}

See the Examples section below.
}
\details{
The function will determine which columns are predictors and outcomes in the
linear model; no user intervention is required.

Missing values are removed for each predictor/outcome combination being
scored.
In cases where \code{\link[=lm]{lm()}} or \code{\link[=anova]{anova()}} fail, the scoring proceeds silently, and
a missing value is given for the score.
}
\examples{
if (rlang::is_installed("modeldata")) {

  # Analysis of variance where `class` is the predictor and the numeric
  # predictors are the responses

  cell_data <- modeldata::cells
  cell_data$case <- NULL

  cell_p_val_res <-
    score_aov_pval |>
    fit(class ~ ., data = cell_data)
  cell_p_val_res@results

  cell_t_stat_res <-
    score_aov_fstat |>
    fit(class ~ ., data = cell_data)
  cell_t_stat_res@results

  # ----------------------------------------------------------------------------
  library(dplyr)

  # Analysis of variance where `chem_fp_*` are the predictors and the numeric
  # `permeability` is the response

  permeability <-
    modeldata::permeability_qsar |>
    # Make the problem a little smaller for time; use 50 predictors
    select(1:51) |>
    # Make the binary predictor columns into factors
    mutate(across(starts_with("chem_fp"), as.factor))

  perm_p_val_res <-
    score_aov_pval |>
    fit(permeability ~ ., data = permeability)
  perm_p_val_res@results

  # Note that some `lm()` calls failed and are given NA score values. For
  # example:
  table(permeability$chem_fp_0007)

  perm_t_stat_res <-
    score_aov_fstat |>
    fit(permeability ~ ., data = permeability)
  perm_t_stat_res@results
}
}
\keyword{datasets}
