% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-cor.R
\name{score_cor}
\alias{score_cor}
\title{Create a score object for correlation coefficients}
\usage{
score_cor(
  range = c(-1, 1),
  fallback_value = 1,
  score_type = "pearson",
  direction = "maximize"
)
}
\arguments{
\item{range}{A numeric vector of length two, specifying the minimum and maximum
possible values, respectively.}

\item{fallback_value}{A numeric scalar used as a fallback value. One of:
\itemize{
\item \code{1} (default)
}}

\item{score_type}{A character string indicating the type of scoring metric to compute.
One of:
\itemize{
\item \code{"pearson"} (default)
\item \code{"spearman"}
}}

\item{direction}{A character string indicating the optimization direction. One of:
\itemize{
\item \code{"maximize"} (default)
\item \code{"minimize"}
\item \code{"target"}
}}

\item{outcome_type}{A character string indicating the outcome type. One of:
\itemize{
\item \code{"numeric"}
}}

\item{predictor_type}{A character string indicating the predictor type. One of:
\itemize{
\item \code{"numeric"}
}}
}
\value{
A score object containing associated metadata such as \code{range}, \code{fallback_value},
\code{score_type} (\code{"pearson"} or \code{"spearman"}), \code{direction}, and other relevant attributes.
}
\description{
Construct a score object containing metadata for univariate feature scoring using the
correlation coefficients.
Output a score object containing associated metadata such as \code{range}, \code{fallback_value},
\code{score_type} (\code{"pearson"} or \code{"spearman"}), \code{direction}, and other relevant attributes.
}
\examples{
# Create a score object
score_cor()
# Change score type
score_cor(score_type = "spearman")
}
