% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-cor.R
\docType{data}
\name{score_cor_pearson}
\alias{score_cor_pearson}
\alias{score_cor_spearman}
\title{Scoring via correlation coefficient}
\format{
An object of class \code{filtro::class_score_cor} (inherits from \code{filtro::class_score}, \code{S7_object}) of length 1.

An object of class \code{filtro::class_score_cor} (inherits from \code{filtro::class_score}, \code{S7_object}) of length 1.
}
\usage{
score_cor_pearson

score_cor_spearman
}
\description{
These objects are used when:
\itemize{
\item The predictors are numeric and the outcome is numeric.
}

In this case, a correlation coefficient (via \code{\link[stats:cor]{stats::cor()}}) is computed with the proper
variable roles.

\code{score_cor_pearson} and \code{score_cor_spearman} are objects that define the technique.
To apply the filter on data, you would use the \code{\link[=fit]{fit()}} method:

\preformatted{
  fit(score_cor_pearson, formula, data)
}

See the Examples section below.
}
\keyword{datasets}
