% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-cross_tab.R
\name{score_cross_tab}
\alias{score_cross_tab}
\title{Create a score object for cross tabulation p-values}
\usage{
score_cross_tab(
  range = c(0, 1),
  trans = NULL,
  score_type = "pval_chisq",
  direction = "minimize"
)
}
\arguments{
\item{range}{NULL}

\item{trans}{NULL}

\item{score_type}{NULL}

\item{direction}{NULL}
}
\description{
Create a score object for cross tabulation p-values
}
\examples{
NULL
}
