% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-forest_imp.R
\name{score_forest_imp}
\alias{score_forest_imp}
\title{Create score object for feature importance scores}
\usage{
score_forest_imp(
  range = c(0, Inf),
  trans = NULL,
  score_type = "imp_rf",
  direction = "maximize"
)
}
\arguments{
\item{range}{NULL}

\item{trans}{NULL}

\item{score_type}{NULL}

\item{direction}{NULL}
}
\description{
Create score object for feature importance scores
}
\examples{
NULL
}
