% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{score_list}
\alias{score_list}
\title{Construct an S7 subclass of base R's \code{list}}
\format{
An object of class \code{S7_S3_class} of length 3.
}
\usage{
score_list
}
\description{
Output an S7 subclass of S3 base R's \code{list}, used in method dispatch for \code{\link[=bind_scores]{bind_scores()}}.
}
\keyword{datasets}
