% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-roc_auc.R
\name{score_roc_auc}
\alias{score_roc_auc}
\title{Create score object for area under the Receiver Operating Characteristic curve (ROC AUC)}
\usage{
score_roc_auc(range = c(0, 1), trans = NULL)
}
\arguments{
\item{range}{NULL}

\item{trans}{NULL}
}
\description{
Create score object for area under the Receiver Operating Characteristic curve (ROC AUC)
}
\examples{
NULL
}
