% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-cross_tab.R
\docType{data}
\name{score_xtab_pval_chisq}
\alias{score_xtab_pval_chisq}
\alias{score_xtab_pval_fisher}
\title{Scoring via the chi-squared test or Fisher’s exact test}
\format{
An object of class \code{filtro::class_score_xtab} (inherits from \code{filtro::class_score}, \code{S7_object}) of length 1.

An object of class \code{filtro::class_score_xtab} (inherits from \code{filtro::class_score}, \code{S7_object}) of length 1.
}
\usage{
score_xtab_pval_chisq

score_xtab_pval_fisher
}
\arguments{
\item{object}{A score class object based on \code{class_score_xtab}.}

\item{formula}{A standard R formula with a single outcome on the right-hand
side and one or more predictors (or \code{.}) on the left-hand side. The data are
processed via \code{\link[stats:model.frame]{stats::model.frame()}}.}

\item{data}{A data frame containing the relevant columns defined by the
formula.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
These objects are used when:
\itemize{
\item The predictors are factors and the outcome is a factor.
}

In this case, a contingency table (via \code{\link[=table]{table()}}) is created with the proper
variable roles, and the cross tabulation p-value is computed using either
the chi-squared test (via \code{\link[stats:chisq.test]{stats::chisq.test()}}) or Fisher's exact test
(via \code{\link[stats:fisher.test]{stats::fisher.test()}}). The p-value that is returned is transformed to
be \code{-log10(p_value)} so that larger values are associated with more important
predictors.

\code{score_xtab_pval_chisq} and \code{score_xtab_pval_fisher} are objects that define the technique.
To apply the filter on data, you would use the \code{\link[=fit]{fit()}} method:

\preformatted{
  fit(score_xtab_pval_chisq, formula, data)
}

See the Examples section below.
}
\details{
The function will determine which columns are predictors and outcomes in the
contingency table; no user intervention is required.

Missing values are removed for each predictor/outcome combination being
scored.
In cases where \code{\link[stats:chisq.test]{stats::chisq.test()}} or \code{\link[stats:fisher.test]{stats::fisher.test()}} fail, the scoring proceeds
silently, and a missing value is given for the score.
}
\examples{
if (rlang::is_installed("titanic")) {

  # Binary factor example

  library(dplyr)

  titanic_subset <- titanic_train |>
    mutate(across(c(Survived, Pclass, Sex, Embarked), as.factor)) |>
    select(Survived, Pclass, Sex, Age, Fare, Embarked)

  titanic_xtab_pval_chisq_res <- score_xtab_pval_chisq |>
    fit(Survived ~ ., data = titanic_subset)
  titanic_xtab_pval_chisq_res@results

  titanic_xtab_pval_fisher_res <- score_xtab_pval_fisher |>
    fit(Survived ~ ., data = titanic_subset)
  titanic_xtab_pval_fisher_res@results
}
  # TODO Add multiclass factor example
}
\keyword{datasets}
