% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desirability2.R
\name{show_best_desirability_num}
\alias{show_best_desirability_num}
\title{Show best desirability scores, based on number of predictors \emph{(plural)}}
\usage{
show_best_desirability_num(x, ..., num_terms = 5)
}
\arguments{
\item{x}{A tibble or data frame returned by \code{\link[=fill_safe_values]{fill_safe_values()}}.}

\item{...}{NULL}

\item{num_terms}{A numeric value specifying the number
of predictors to consider.}
}
\value{
Returns a tibble with \code{num_terms}
number of rows. When showing the results,
the metrics are presented in "wide format" (one column per metric) and there
are new columns for the corresponding desirability values (each starts with
\code{.d_}).
}
\description{
Similar to \code{\link[=show_best_desirability_prop]{show_best_desirability_prop()}} that can
simultaneously optimize multiple scores using desirability functions.
See \code{\link[=show_best_score_num]{show_best_score_num()}} for \emph{singular} scoring method.
}
\details{
See \code{\link[=show_best_desirability_prop]{show_best_desirability_prop()}} for details.
}
\seealso{
\code{\link[=show_best_desirability_prop]{show_best_desirability_prop()}}, \code{\link[=show_best_desirability_cutoff]{show_best_desirability_cutoff()}}
}
