% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{show_best_score_cutoff}
\alias{show_best_score_cutoff}
\title{Show best score, based on based on cutoff value \emph{(singular)}}
\usage{
show_best_score_cutoff(x, ...)
}
\arguments{
\item{x}{A score class object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Show best score, based on based on cutoff value \emph{(singular)}
}
