% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{show_best_score_num}
\alias{show_best_score_num}
\title{Show best score, based on number of predictors \emph{(singular)}}
\arguments{
\item{x}{A score class object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Show best score, based on number of predictors \emph{(singular)}
}
