% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{show_best_score_prop}
\alias{show_best_score_prop}
\title{Show best score based on proportion of predictors}
\usage{
show_best_score_prop(x, ...)
}
\arguments{
\item{x}{A score class object.}

\item{...}{NULL}
}
\description{
Show best score based on proportion of predictors
}
