% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{trans_score}
\alias{trans_score}
\title{Transform score result \code{score_res}}
\usage{
trans_score(x, ...)
}
\arguments{
\item{x}{NULL}

\item{...}{NULL}
}
\description{
Transform score result \code{score_res}
}
