% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{transform_abs}
\alias{transform_abs}
\title{Absolute Transformation}
\format{
An object of class \code{transform} of length 9.
}
\usage{
transform_abs
}
\description{
Absolute Transformation
}
\keyword{datasets}
