% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{transform_neg_log10}
\alias{transform_neg_log10}
\title{Negative log10 Transformation (TODO Rewrite using other method)}
\format{
An object of class \code{transform} of length 9.
}
\usage{
transform_neg_log10
}
\description{
Negative log10 Transformation (TODO Rewrite using other method)
}
\keyword{datasets}
