% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{transform_score}
\alias{transform_score}
\title{Transform score \emph{(plural)}}
\arguments{
\item{x}{A tibble or data frame returned by \code{\link[=fill_safe_values]{fill_safe_values()}}.}

\item{trans}{A named list of functions.}
}
\value{
A tibble of scores results.
}
\description{
Transform score \emph{(plural)}
}
