% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ames_scores_results}
\alias{ames_scores_results}
\title{Ames exampled score results}
\value{
\item{ames_scores_results}{a tibble}
}
\description{
This data an ames exampled score results for, used
for \code{\link[=show_best_desirability_prop]{show_best_desirability_prop()}} as a demonstration,
and created by examples in \code{\link[=fill_safe_values]{fill_safe_values()}}.
}
\examples{
data(ames_scores_results)
}
