% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{arrange_score}
\alias{arrange_score}
\title{Arrange score to filter object}
\usage{
arrange_score(score_obj, ...)
}
\arguments{
\item{score_obj}{NULL}

\item{...}{NULL}
}
\description{
Arrange score to filter object
}
