% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{arrange_score}
\alias{arrange_score}
\title{Arrange score}
\arguments{
\item{x}{A score class object, i.e., \verb{score_*}.}

\item{...}{Further arguments passed to or from other methods.}

\item{target}{A numeric value specifying the target value. The default
of \code{NULL} indicates that there is no target value.}
}
\description{
Arrange score
}
\examples{
\dontshow{if (rlang::is_installed("modeldata")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(dplyr)

ames <- modeldata::ames
ames <- ames |>
  dplyr::mutate(Sale_Price = log10(Sale_Price))

ames_aov_pval_res <-
  score_aov_pval |>
  fit(Sale_Price ~ ., data = ames)
ames_aov_pval_res@results

# Arrange score
ames_aov_pval_res |> arrange_score()
\dontshow{\}) # examplesIf}
}
