% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{arrange_score}
\alias{arrange_score}
\title{Arrange score}
\arguments{
\item{x}{A score class object (e.g., \code{score_cor_pearson}).}

\item{...}{Further arguments passed to or from other methods.}

\item{target}{A numeric value specifying the target value. The default
of \code{NULL} indicates that there is no target value.}
}
\value{
A tibble of score results.
}
\description{
Arrange score
}
\examples{
\dontshow{if (rlang::is_installed("modeldata")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(dplyr)

ames_subset <- modeldata::ames |>
  dplyr::select(
    Sale_Price,
    MS_SubClass,
    MS_Zoning,
    Lot_Frontage,
    Lot_Area,
    Street
  )
ames_subset <- ames_subset |>
  dplyr::mutate(Sale_Price = log10(Sale_Price))

ames_aov_pval_res <-
  score_aov_pval |>
  fit(Sale_Price ~ ., data = ames_subset)
ames_aov_pval_res@results

# Arrange score
ames_aov_pval_res |> arrange_score()
\dontshow{\}) # examplesIf}
}
