% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{attach_score}
\alias{attach_score}
\title{Attach score result \code{results} to score object \code{score_obj} containig metadata}
\usage{
attach_score(x, results, ...)
}
\arguments{
\item{x}{NULL}

\item{results}{NULL}

\item{...}{NULL}
}
\description{
Attach score result \code{results} to score object \code{score_obj} containig metadata
}
