% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{attach_score}
\alias{attach_score}
\title{Attach score to filter object}
\usage{
attach_score(x, ...)
}
\arguments{
\item{...}{NULL}

\item{score_obj}{NULL}
}
\description{
Attach score to filter object
}
