% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bind_scores}
\alias{bind_scores}
\title{Bind scores with metadata \code{score_obj} and results \code{res}, and assign class \code{score_set}.}
\usage{
bind_scores(x)
}
\arguments{
\item{x}{NULL}
}
\description{
Bind scores with metadata \code{score_obj} and results \code{res}, and assign class \code{score_set}.
}
