% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bind_scores}
\alias{bind_scores}
\title{Bind all metadata \code{score_obj} and score \code{res}, and assign class \code{score_set} to scores.}
\usage{
bind_scores(x)
}
\arguments{
\item{x}{NULL}
}
\description{
Bind all metadata \code{score_obj} and score \code{res}, and assign class \code{score_set} to scores.
}
