% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fill_safe_value}
\alias{fill_safe_value}
\title{Fill safe value}
\usage{
fill_safe_value(x, ...)
}
\arguments{
\item{x}{A score class object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Fill safe value
}
\examples{
\dontshow{if (rlang::is_installed("modeldata")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(dplyr)

ames_subset <- modeldata::ames |>
  dplyr::select(
    Sale_Price,
    MS_SubClass,
    MS_Zoning,
    Lot_Frontage,
    Lot_Area,
    Street
  )
ames_subset <- ames_subset |>
  dplyr::mutate(Sale_Price = log10(Sale_Price))
\dontshow{\}) # examplesIf}
}
