% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fill_safe_values}
\alias{fill_safe_values}
\title{Fill in safe values.}
\usage{
fill_safe_values(x)
}
\arguments{
\item{x}{A list where each element is a score object of class \code{score_obj}.}
}
\description{
Fill in safe values.
}
