% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filter_score_cutoff}
\alias{filter_score_cutoff}
\title{Filter score result \code{results} based on cutoff value}
\usage{
filter_score_cutoff(x, ...)
}
\arguments{
\item{x}{NULL}

\item{...}{NULL}
}
\description{
Filter score result \code{results} based on cutoff value
}
