% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtro-package.R
\docType{package}
\name{filtro-package}
\alias{filtro}
\alias{filtro-package}
\title{filtro: Feature Selection Using Supervised Filter-Based Methods}
\description{
Tidy tools to apply filter-based supervised feature selection methods. These methods score and rank feature relevance using metrics such as p-values, correlation, and importance scores (Kuhn and Johnson (2019) \doi{10.1201/9781315108230}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/filtro}
  \item Report bugs at \url{https://github.com/tidymodels/filtro/issues}
}

}
\author{
\strong{Maintainer}: Frances Lin \email{franceslinyc@gmail.com}

Authors:
\itemize{
  \item Max Kuhn \email{max@posit.co}
  \item Emil Hvitfeldt \email{emil.hvitfeldt@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
