% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-aov.R
\name{get_scores_aov}
\alias{get_scores_aov}
\title{Compute F-statistic and p-value scores using ANOVA F-test}
\usage{
get_scores_aov(score_obj, data, outcome)
}
\arguments{
\item{score_obj}{A score object. See \code{\link[=score_aov]{score_aov()}} for details.}

\item{data}{A data frame or tibble containing the outcome and predictor variables.}

\item{outcome}{A character string specifying the name of the outcome variable.}

\item{neg_log10}{A logical value, indicating whether to apply a negative log10
transformation to p-values (\code{TRUE}) or not (\code{FALSE}).
If \code{neg_log10} is \code{TRUE}, the transformation is applied, and \code{direction} should be
set to \code{"maximize"}. If \code{neg_log10} is \code{FALSE}, raw p-values are used, and \code{direction}
should be set to \code{"minimize"}. By default, \code{neg_log10} is set to \code{TRUE}.}
}
\value{
A tibble of result with one row per predictor, and four columns:
\itemize{
\item \code{name}: the name of scoring metric.
\item \code{score}: the score for the predictor-outcome pair.
\item \code{predictor}: the name of the predictor.
\item \code{outcome}: the name of the outcome.
}
}
\description{
Output a tibble result with with one row per predictor, and four columns:
\code{name}, \code{score}, \code{predictor}, and \code{outcome}.
}
\examples{
data(ames, package = "modeldata")
data <- modeldata::ames |>
  dplyr::select(
    Sale_Price,
    MS_SubClass,
    MS_Zoning,
    Lot_Frontage,
    Lot_Area,
    Street
  )
outcome <- "Sale_Price"
score_obj <- score_aov()
score_res <- get_scores_aov(score_obj, data, outcome)
score_res
}
