% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-cor.R
\name{get_scores_cor}
\alias{get_scores_cor}
\title{Compute Pearson or Spearman correlation coefficients}
\usage{
get_scores_cor(score_obj, data, outcome)
}
\arguments{
\item{score_obj}{A score object. See \code{\link[=score_cor]{score_cor()}} for details.}

\item{data}{A data frame or tibble containing the outcome and predictor variables.}

\item{outcome}{A character string specifying the name of the outcome variable.}
}
\value{
A tibble of result with one row per predictor, and four columns:
\itemize{
\item \code{name}: the name of scoring metric.
\item \code{score}: the score for the predictor-outcome pair.
\item \code{predictor}: the name of the predictor.
\item \code{outcome}: the name of the outcome.
}
}
\description{
Evaluate the relationship between a numeric outcome and a numeric predictor,
by computing the Pearson or Spearman correlation coefficients.
Output a tibble result with with one row per predictor, and four columns:
\code{name}, \code{score}, \code{predictor}, and \code{outcome}.
}
\examples{
data(ames, package = "modeldata")
ames_subset <- modeldata::ames |>
  dplyr::select(
    Sale_Price,
    MS_SubClass,
    MS_Zoning,
    Lot_Frontage,
    Lot_Area,
    Street
  )
# Return score as pearson correlation
score_obj <- score_cor()
score_res <- get_scores_cor(
  score_obj,
  data = ames_subset,
  outcome = "Sale_Price"
)
score_res
# Return score as spearman correlation
score_obj <- score_cor(score_type = "spearman")
score_res <- get_scores_cor(
  score_obj,
  data = ames_subset,
  outcome = "Sale_Price"
)
score_res
}
