% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-roc_auc.R
\name{get_scores_roc_auc}
\alias{get_scores_roc_auc}
\title{Compute area under the Receiver Operating Characteristic curve (ROC AUC)}
\usage{
get_scores_roc_auc(score_obj, data, outcome, ...)
}
\arguments{
\item{score_obj}{A score object. See \code{\link[=score_roc_auc]{score_roc_auc()}} for details.}

\item{data}{A data frame or tibble containing the outcome and predictor variables.}

\item{outcome}{A character string specifying the name of the outcome variable.}

\item{...}{NULL}
}
\value{
A tibble of result with one row per predictor, and four columns:
\itemize{
\item \code{name}: the name of scoring metric.
\item \code{score}: the score for the predictor-outcome pair.
\item \code{predictor}: the name of the predictor.
\item \code{outcome}: the name of the outcome.
}
}
\description{
Evaluate the relationship between a numeric outcome and a categorical predictor,
or vice versa, by computing the area under the Receiver Operating Characteristic curve (ROC AUC).
Output a tibble result with with one row per predictor, and four columns:
\code{name}, \code{score}, \code{predictor}, and \code{outcome}.
}
\examples{
cells_subset <- modeldata::cells |>
  dplyr::select(
    class,
    angle_ch_1,
    area_ch_1,
    avg_inten_ch_1,
    avg_inten_ch_2
  )
# Return score as ROC AUC for binary classification
score_obj = score_roc_auc()
score_res <- get_scores_roc_auc(
  score_obj,
  data = cells_subset,
  outcome = "class"
)
# Return score as ROC AUC for multiclass classification
}
