% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{new_score_obj}
\alias{new_score_obj}
\title{Title}
\usage{
new_score_obj(
  subclass = c("cat_num", "cat_cat", "num_num", "any"),
  outcome_type = c("numeric", "factor"),
  predictor_type = c("numeric", "factor"),
  case_weights = NULL,
  range = NULL,
  inclusive = NULL,
  fallback_value = NULL,
  score_type = NULL,
  trans = NULL,
  sorts = NULL,
  direction = NULL,
  deterministic = NULL,
  tuning = NULL,
  ties = NULL,
  calculating_fn = NULL,
  label = NULL,
  ...
)
}
\arguments{
\item{subclass}{NULL}

\item{outcome_type}{NULL}

\item{predictor_type}{NULL}

\item{case_weights}{NULL}

\item{range}{NULL}

\item{inclusive}{NULL}

\item{fallback_value}{NULL}

\item{score_type}{NULL}

\item{trans}{NULL}

\item{sorts}{NULL}

\item{direction}{NULL}

\item{deterministic}{NULL}

\item{tuning}{NULL}

\item{ties}{NULL}

\item{calculating_fn}{NULL}

\item{label}{NULL}

\item{...}{}
}
\description{
Title
}
\examples{
NULL
}
