% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rank_best_score_dense}
\alias{rank_best_score_dense}
\title{Rank score based on \code{dplyr::dense_rank()}, where tied values receive the
same rank and ranks are with gaps \emph{(singular)}}
\usage{
rank_best_score_dense(x, ...)
}
\arguments{
\item{x}{A score class object.}

\item{...}{NULL}
}
\description{
Rank score based on \code{dplyr::dense_rank()}, where tied values receive the
same rank and ranks are with gaps \emph{(singular)}
}
