% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rank_best_score_min}
\alias{rank_best_score_min}
\title{Rank score based on min_rank(), where tied values receive the same (smalles) rank
and ranks are with gaps}
\usage{
rank_best_score_min(x, ...)
}
\arguments{
\item{x}{A score class object.}

\item{...}{NULL}
}
\description{
Rank score based on min_rank(), where tied values receive the same (smalles) rank
and ranks are with gaps
}
