% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-aov.R
\docType{data}
\name{score_aov}
\alias{score_aov}
\alias{score_aov_pval}
\alias{score_aov_fstat}
\title{Scoring via analysis of variance hypothesis tests}
\format{
An object of class \code{filtro::class_score_aov} (inherits from \code{filtro::class_score}, \code{S7_object}) of length 1.

An object of class \code{filtro::class_score_aov} (inherits from \code{filtro::class_score}, \code{S7_object}) of length 1.
}
\usage{
score_aov_pval

score_aov_fstat
}
\description{
This method is used when either:
\itemize{
\item The predictors are numeric and the outcome is a factor/category, or
\item The predictors are factors and the outcome is numeric.
}

In either case, a linear model (vi \code{\link[=lm]{lm()}}) is created with the proper
variable roles, and the overall p-value for the hypothesis that all means are
equal is computed via the standard F-statistic. The p-value that is returned
is transformed to be \code{-log10(p_value)} so that larger values are associated
with more important predictors.
}
\keyword{datasets}
