% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-cor.R
\name{score_cor}
\alias{score_cor}
\title{Compute Pearson or Spearman correlation coefficients}
\usage{
score_cor(
  range = c(-1, 1),
  trans = NULL,
  score_type = "pearson",
  direction = "maximize"
)
}
\arguments{
\item{range}{NULL}

\item{trans}{NULL}

\item{score_type}{NULL}

\item{direction}{NULL}
}
\description{
Compute Pearson or Spearman correlation coefficients
}
\examples{
NULL
}
