% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-forest_imp.R
\docType{data}
\name{score_imp_rf}
\alias{score_imp_rf}
\alias{score_imp_rf_conditional}
\alias{score_imp_rf_oblique}
\title{Scoring via random forest, conditional random forest, or oblique random forest}
\format{
An object of class \code{filtro::class_score_imp_rf} (inherits from \code{filtro::class_score}, \code{S7_object}) of length 1.

An object of class \code{filtro::class_score_imp_rf} (inherits from \code{filtro::class_score}, \code{S7_object}) of length 1.

An object of class \code{filtro::class_score_imp_rf} (inherits from \code{filtro::class_score}, \code{S7_object}) of length 1.
}
\usage{
score_imp_rf

score_imp_rf_conditional

score_imp_rf_oblique
}
\description{
These objects are used when either:
\itemize{
\item The predictors are numeric and the outcome is a factor/category, or
\item The predictors are factors and the outcome is numeric.
}

In either case, a random forest (via \code{\link[ranger:ranger]{ranger::ranger()}}, \code{\link[partykit:cforest]{partykit::cforest()}},
or \code{\link[aorsf:orsf]{aorsf::orsf()}}) is created with the proper variable roles, and the feature
importance scores are computed. The larger values are associated with more
important predictors.

\code{score_imp_rf}, \code{score_imp_rf_conditional} and \code{score_imp_rf_oblique} are
objects that define the technique.
To apply the filter on data, you would use the \code{\link[=fit]{fit()}} method:

\preformatted{
  fit(score_imp_rf, formula, data)
}

See the Examples section below.
}
\keyword{datasets}
