% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-info_gain.R
\name{score_info_gain}
\alias{score_info_gain}
\title{Create score object for information gain}
\usage{
score_info_gain(
  range = c(0, Inf),
  trans = NULL,
  score_type = "infogain",
  direction = "maximize"
)
}
\arguments{
\item{range}{NULL}

\item{trans}{NULL}

\item{score_type}{NULL}

\item{direction}{NULL}
}
\description{
Create score object for information gain
}
\examples{
NULL
}
