% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-info_gain.R
\name{score_info_gain}
\alias{score_info_gain}
\title{Create a score object for information gain}
\usage{
score_info_gain(
  range = c(0, Inf),
  fallback_value = Inf,
  score_type = "infogain",
  direction = "maximize",
  is_reg = FALSE
)
}
\arguments{
\item{range}{A numeric vector of length two, specifying the minimum and maximum
possible values, respectively.}

\item{fallback_value}{A numeric scalar used as a fallback value. One of:
\itemize{
\item \code{1}
\item \code{Inf} (default)
}}

\item{score_type}{A character string indicating the type of scoring metric to compute.
One of:
\itemize{
\item \code{"infogain"} (default)
\item \code{"gainratio"}
\item \code{"symuncert"}
}}

\item{direction}{A character string indicating the optimization direction. One of:
\itemize{
\item \code{"maximize"} (default)
\item \code{"minimize"}
\item \code{"target"}
}}

\item{is_reg}{NULL}
}
\value{
A score object containing associated metadata such as \code{range}, \code{fallback_value},
\code{score_type} (), \code{direction}, and other relevant attributes.
}
\description{
Construct a score object containing metadata for feature scoring using
information gain.
Output a score object containing associated metadata such as \code{range}, \code{fallback_value},
\code{score_type} (), \code{direction}, and other relevant attributes.
}
\examples{
score_info_gain()
}
