% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-roc_auc.R
\docType{data}
\name{score_roc_auc}
\alias{score_roc_auc}
\title{Scoring via area under the Receiver Operating Characteristic curve (ROC AUC)}
\format{
An object of class \code{filtro::class_score_roc_auc} (inherits from \code{filtro::class_score}, \code{S7_object}) of length 1.
}
\usage{
score_roc_auc
}
\arguments{
\item{object}{A score class object based on \code{class_score_roc_auc}.}

\item{formula}{A standard R formula with a single outcome on the right-hand
side and one or more predictors (or \code{.}) on the left-hand side. The data are
processed via \code{\link[stats:model.frame]{stats::model.frame()}}.}

\item{data}{A data frame containing the relevant columns defined by the
formula.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
These objects are used when either:
\itemize{
\item The predictors are numeric and the outcome is a factor/category, or
\item The predictors are factors and the outcome is numeric.
}

In either case, a ROC curve (via \code{\link[pROC:roc]{pROC::roc()}}) is created with the proper
variable roles, and the area under the ROC curve is computed (via \code{\link[pROC:auc]{pROC::auc()}}).
Values higher than 0.5 (i.e., \code{max(roc_auc, 1 - roc_auc)} > 0.5) are associated with
more important predictors.

\code{score_roc_auc} is object that define the technique.
To apply the filter on data, you would use the \code{\link[=fit]{fit()}} method:

\preformatted{
  fit(score_roc_auc, formula, data)
}

See the Examples section below.
}
\details{
The function will determine which columns are predictors and outcomes for the
ROC analysis; no user intervention is required.

Missing values are removed for each predictor/outcome combination being
scored.
In cases where \code{\link[pROC:roc]{pROC::roc()}} fail, the scoring proceeds silently, and
a missing value is given for the score.
}
\examples{
if (rlang::is_installed("modeldata")) {

  library(dplyr)

  # ROC AUC where the numeric predictors are the predictors and
  # `class` is the class outcome/response

  cells_subset <- modeldata::cells |>
    dplyr::select(
      class,
      angle_ch_1,
      area_ch_1,
      avg_inten_ch_1,
      avg_inten_ch_2,
      avg_inten_ch_3
    )

  cells_roc_auc_res <- score_roc_auc |>
    fit(class ~ ., data = cells_subset)
  cells_roc_auc_res@results

  # ----------------------------------------------------------------------------

  # ROC AUC where `Sale_Price` is the numeric predictor and the class predictors
  # are the outcomes/responses

  ames_subset <- modeldata::ames |>
    dplyr::select(
      Sale_Price,
      MS_SubClass,
      MS_Zoning,
      Lot_Frontage,
      Lot_Area,
      Street
    )
  ames_subset <- ames_subset |>
    dplyr::mutate(Sale_Price = log10(Sale_Price))

  ames_roc_auc_res <- score_roc_auc |>
    fit(Sale_Price ~ ., data = ames_subset)
  ames_roc_auc_res@results
}
  # TODO Add multiclass example
}
\keyword{datasets}
