% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desirability2.R
\name{show_best_desirability_prop}
\alias{show_best_desirability_prop}
\title{Show best desirability scores, based on proportion of predictors \emph{(plural)}}
\usage{
show_best_desirability_prop(x, ..., prop_terms = 0.99)
}
\arguments{
\item{x}{A tibble. The results of \code{\link[=fill_safe_values]{fill_safe_values()}}.}

\item{...}{NULL}

\item{prop_terms}{A numeric value specifying the proportion
of predictors to consider.}
}
\value{
\code{\link[=show_best_desirability_prop]{show_best_desirability_prop()}} returns a tibble with \code{prop_terms}
proportion of rows. When showing the results,
the metrics are presented in "wide format" (one column per metric) and there
are new columns for the corresponding desirability values (each starts with
\code{.d_}).
}
\description{
Analogous to, and adapted from \code{\link[desirability2:show_best_desirability]{desirability2::show_best_desirability()}} that can
simultaneously optimize multiple scores using desirability functions.
See also \code{\link[=show_best_score_prop]{show_best_score_prop()}} for \emph{singular} scoring method.
}
