% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{show_best_score_dual}
\alias{show_best_score_dual}
\title{Show best score based on number or proportion of predictors with
optional cutoff value}
\usage{
show_best_score_dual(x, ...)
}
\arguments{
\item{x}{A score class object.}

\item{...}{NULL}
}
\description{
Show best score based on number or proportion of predictors with
optional cutoff value
}
