% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{show_best_score_prop}
\alias{show_best_score_prop}
\title{Show best score, based on proportion of predictors \emph{(singular)}}
\arguments{
\item{x}{A score class object.}

\item{...}{NULL}

\item{prop_terms}{A numeric value specifying the proportion
of predictors to consider.}
}
\description{
Show best score, based on proportion of predictors \emph{(singular)}
}
