% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{trans_score}
\alias{trans_score}
\title{Transform score to score object}
\usage{
trans_score(score_obj, ...)
}
\arguments{
\item{score_obj}{NULL}

\item{...}{NULL}
}
\description{
Transform score to score object
}
